requirements.txt
Flask
Jinja2
docopt
WTForms
flask-paginate
Flask-SQLAlchemy
Flask-Migrate